% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{CountryPopulation
This function takes a country name as the argument and returns a graph
showing the country's population by year}
\usage{
CountryPopulation(country_name)
}
\arguments{
\item{country_name}{A string}
}
\value{
A graph displaying the country's population  by year
}
\description{
CountryPopulation
This function takes a country name as the argument and returns a graph
showing the country's population by year
}
\examples{
CountryPopulation('China')
CountryPopulation('France')
}
