% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{World_Cup}
\format{

}
\usage{
World_Cup
}
\description{
This script reads the raw data the Wikipedia World Cup page,
selects the columns Year, Hosts, Matches, Total_Attendance,
Average_Attendance. Then it saves it as an rda file in the data directory.
}
\keyword{datasets}
